+++ contracts/ibc-reflect/src/contract.rs
@@ -1,15 +1,15 @@
 use cosmwasm_std::{
     entry_point, from_json, to_json_binary, wasm_execute, BankMsg, Binary, CosmosMsg, Deps,
-    DepsMut, Empty, Env, Event, Ibc3ChannelOpenResponse, IbcBasicResponse, IbcChannelCloseMsg,
-    IbcChannelConnectMsg, IbcChannelOpenMsg, IbcChannelOpenResponse, IbcOrder, IbcPacketAckMsg,
-    IbcPacketReceiveMsg, IbcPacketTimeoutMsg, IbcReceiveResponse, MessageInfo, Never,
-    QueryResponse, Reply, Response, StdError, StdResult, SubMsg, SubMsgResponse, SubMsgResult,
-    WasmMsg,
+    DepsMut, Empty, Env, Event, Ibc3ChannelOpenResponse, IbcAcknowledgement, IbcBasicResponse,
+    IbcChannelCloseMsg, IbcChannelConnectMsg, IbcChannelOpenMsg, IbcChannelOpenResponse, IbcMsg,
+    IbcOrder, IbcPacketAckMsg, IbcPacketReceiveMsg, IbcPacketTimeoutMsg, IbcReceiveResponse,
+    MessageInfo, Never, QueryResponse, Reply, Response, StdError, StdResult, SubMsg,
+    SubMsgResponse, SubMsgResult, WasmMsg,
 };
 
 use crate::msg::{
     AccountInfo, AccountResponse, AcknowledgementMsg, BalancesResponse, DispatchResponse,
-    InstantiateMsg, ListAccountsResponse, PacketMsg, QueryMsg, ReflectExecuteMsg,
+    ExecuteMsg, InstantiateMsg, ListAccountsResponse, PacketMsg, QueryMsg, ReflectExecuteMsg,
     ReturnMsgsResponse, WhoAmIResponse,
 };
 use crate::state::{
@@ -35,6 +35,34 @@ pub fn instantiate(
     save_item(deps.storage, KEY_CONFIG, &cfg)?;
 
     Ok(Response::new().add_attribute("action", "instantiate"))
+}
+
+#[entry_point]
+pub fn execute(
+    _deps: DepsMut,
+    _env: Env,
+    _info: MessageInfo,
+    msg: ExecuteMsg,
+) -> StdResult<Response> {
+    match msg {
+        ExecuteMsg::AsyncAck {
+            channel_id,
+            packet_sequence,
+            ack,
+        } => execute_async_ack(channel_id, packet_sequence.u64(), ack),
+    }
+}
+
+fn execute_async_ack(
+    channel_id: String,
+    packet_sequence: u64,
+    ack: IbcAcknowledgement,
+) -> StdResult<Response> {
+    Ok(Response::new().add_message(IbcMsg::WriteAcknowledgement {
+        channel_id,
+        packet_sequence,
+        ack,
+    }))
 }
 
 #[entry_point]
@@ -249,6 +277,7 @@ pub fn ibc_packet_receive(
             PacketMsg::Panic {} => execute_panic(),
             PacketMsg::ReturnErr { text } => execute_error(text),
             PacketMsg::ReturnMsgs { msgs } => execute_return_msgs(msgs),
+            PacketMsg::NoAck {} => Ok(IbcReceiveResponse::without_ack()),
         }
     })()
     .or_else(|e| {
@@ -607,7 +636,7 @@ mod tests {
         // acknowledgement is an error
         let ack: AcknowledgementMsg<DispatchResponse> =
             from_json(res.acknowledgement.unwrap()).unwrap();
-        assert_eq!(ack.unwrap_err(), "invalid packet: Error parsing into type ibc_reflect::msg::PacketMsg: unknown variant `reflect_code_id`, expected one of `dispatch`, `who_am_i`, `balances`, `panic`, `return_err`, `return_msgs`");
+        assert_eq!(ack.unwrap_err(), "invalid packet: Error parsing into type ibc_reflect::msg::PacketMsg: unknown variant `reflect_code_id`, expected one of `dispatch`, `who_am_i`, `balances`, `panic`, `return_err`, `return_msgs`, `no_ack`");
     }
 
     #[test]